marcar gol del otro equipo (por consola)

+++ server/proyecto-fin-de-grado/src/main/java/com/ikerveiga/app/entity/TeamStats.java
@@ -195,7 +195,7 @@ public class TeamStats {
         return this.penaltyKillPercentage;
     }
 
-    public void setPenaltyKillPercentage(int penaltyKillPercentage) {
+    public void setPenaltyKillPercentage(float penaltyKillPercentage) {
         this.penaltyKillPercentage = penaltyKillPercentage;
     }
